+++ apps/users/function.py
@@ -4,7 +4,7 @@ from fastapi import HTTPException
 from fastapi.encoders import jsonable_encoder
 from sqlalchemy.orm import Session
 from book_management.core.hash import hash_password, verify_password
-from book_management.core.exceptions import reguler_user_exception
+from book_management.core.permission import reguler_user_permission
 
 
 def register_user(db: Session, user: schema.UserRegister):
@@ -59,7 +59,7 @@ def delete_me(db: Session, userid: int):
     return {"message": "account deleted sucesfully"}
 
 
-@reguler_user_exception
+@reguler_user_permission
 def get_all_reader(db: Session, userid: int):
 
     users_obj = crud.get_all_reader(db)
@@ -70,7 +70,7 @@ def get_all_reader(db: Session, userid: int):
                         "error": "There are no readers"})
 
 
-@reguler_user_exception
+@reguler_user_permission
 def get_a_reader(db: Session, user_id: int, reader_id):
     users_obj = crud.get_a_reader(db, reader_id)
     if users_obj is not None:
@@ -80,7 +80,7 @@ def get_a_reader(db: Session, user_id: int, reader_id):
                         "error": "User does not exist"})
 
 
-@reguler_user_exception
+@reguler_user_permission
 def set_status(db, user_id, reader_id, active):
     if crud.set_status(db, reader_id, active):
         return {"message": "user updated sucessfully"}
@@ -89,12 +89,11 @@ def set_status(db, user_id, reader_id, active):
                         "error": "User does not exist"})
 
 
-@reguler_user_exception
+@reguler_user_permission
 def search_reader(db, user_id, filers):
     results = crud.search_reader(db, filers)
     
     if not results:
         raise HTTPException(status_code=404, detail={
                         "error": "resource not found"})
-    
     return jsonable_encoder(results)
